Group triangulation related data into Mesh struct in mesh.rs
