+++ src/agent/find_path.rs
@@ -24,14 +24,12 @@ impl Agent {
         game: &mut Game,
     ) -> Result<(), ()> {
         if let Some(target) = target {
-            let triangulation = &game.triangulation;
-            let points = &game.points;
+            let triangulation = &game.mesh.triangulation;
+            let points = &game.mesh.points;
             let mut profiler = game.triangle_profiler.borrow_mut();
-            let this_triangle =
-                find_triangle_at(triangulation, points, self.pos, &mut *profiler).ok_or(())?;
+            let this_triangle = find_triangle_at(&game.mesh, self.pos, &mut *profiler).ok_or(())?;
             let target_triangle =
-                find_triangle_at(triangulation, points, target.get_pos(), &mut *profiler)
-                    .ok_or(())?;
+                find_triangle_at(&game.mesh, target.get_pos(), &mut *profiler).ok_or(())?;
             if this_triangle == target_triangle {
                 // self.path_line = vec![
                 //         self.pos,
@@ -50,7 +48,7 @@ impl Agent {
                 for j in 0..3 {
                     let next_halfedge = triangulation.halfedges[top * 3 + j];
                     if next_halfedge == delaunator::EMPTY
-                        || !game.triangle_passable[next_halfedge / 3]
+                        || !game.mesh.triangle_passable[next_halfedge / 3]
                     {
                         continue;
                     }
